Try, hope it works
I still have some troubles with python and numpy versions so could not check wheter it works (but fairly simple code so should be fine).  Running nodes.py should yield adj_matrix with distances and weights with demand (change names if you want). Node 1 has demand 0 so I guess it would be the storage point. (Should we allow for a more flexible situation in which the deposit is not in the first position? Or maybe comment somewhere in the documentation to put as first point in the dictionary the deposit).