This min seems better
What I did is to subtract from the max fitness the fitness of each individual. Then I added a quantity so that the least likely individual will have the same probability as it would have in a max problem with the same fitness vector. I debugged it and it does what I intended. I have concerns instead on the overall performance, since it does not seem to improve from initial point (but maybe we need to fine tune parameters first?)